+++ database/migrations/2022_04_27_043404_add_relationships_to_order_details_table.php
@@ -13,17 +13,17 @@ class AddRelationshipsToOrderDetailsTable extends Migration
      */
     public function up()
     {
-        Schema::table('order_details', function(Blueprint $table) {
-        $table->integer('order_id')->unsigned()->change();
-        $table->foreign('order_id')->references('id')->on('orders')
-            ->onUpdate('cascade')
-            ->onDelete('cascade');
+        Schema::table('order_details', function (Blueprint $table) {
+            $table->bigInteger('order_id')->unsigned()->change();
+            $table->foreign('order_id')->references('id')->on('orders')
+                ->onUpdate('cascade')
+                ->onDelete('cascade');
 
-        $table->integer('product_id')->unsigned()->change();
-        $table->foreign('product_id')->references('id')->on('products')
-            ->onUpdate('cascade')
-            ->onDelete('cascade');
-    });
+            $table->bigInteger('product_id')->unsigned()->change();
+            $table->foreign('product_id')->references('id')->on('products')
+                ->onUpdate('cascade')
+                ->onDelete('cascade');
+        });
     }
 
     /**
@@ -33,13 +33,13 @@ class AddRelationshipsToOrderDetailsTable extends Migration
      */
     public function down()
     {
-        Schema::table('order_details', function(Blueprint $table) {
+        Schema::table('order_details', function (Blueprint $table) {
             $table->dropForeign('order_details_order_id_foreign');
             $table->dropIndex('order_details_order_id_foreign');
-            $table->integer('order_id')->change();
+            $table->bigInteger('order_id')->change();
             $table->dropForeign('order_details_product_id_foreign');
             $table->dropIndex('order_details_product_id_foreign');
-            $table->integer('product_id')->change();
+            $table->bigInteger('product_id')->change();
         });
     }
 }
